OPl-9: Added invoice payments

+++ invoice/tests/helpers/invoice_helpers.py
@@ -7,7 +7,7 @@ from policyholder.tests.helpers import create_test_policy_holder
 from insuree.test_helpers import create_test_insuree
 from core.forms import User
 
-from invoice_payment.models import Invoice
+from invoice.models import Invoice
 
 DEFAULT_TEST_INVOICE_PAYLOAD = {
     'subject_type': 'contract',
@@ -24,7 +24,7 @@ DEFAULT_TEST_INVOICE_PAYLOAD = {
     'amount_net': 20.1,
     'tax_analysis': {'lines': [{'code': 'c', 'label': 'l', 'base': '0.1', 'amount': '2.01'}], 'total': '2.01'},
     'amount_total': 20.1,
-    'status': 0,  # Draft
+    'status': Invoice.InvoiceStatus.VALIDATED,  # Validated
     'note': 'NOTE',
     'terms': 'TERMS',
     'payment_reference': 'payment reference'
@@ -47,10 +47,14 @@ def create_test_invoice(subject=None, recipient=None, user=None, **custom_props)
     payload['recipient'] = recipient
     payload['recipient_type'] = ContentType.objects.get_for_model(recipient)
     payload.update(**custom_props)
-
-    Invoice.objects.filter(code=payload['code']).delete()
-
     user = user or __get_or_create_user()
+
+    if Invoice.objects.filter(code=payload['code']).exists():
+        i = Invoice.objects.filter(code=payload['code']).first()
+        i.payments.all().delete()
+        i.line_items.all().delete()
+        Invoice.objects.filter(code=payload['code']).delete()
+
     invoice = Invoice(**payload)
     invoice.save(username=user.username)
